back_movie/requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-cors-headers
drf-writable-nested